// functions/src/types/genkit.ts
export interface QuizItem {
  question: string;
  choices: string[];
  answerIndex: number;
  explanation: string;
  examTopic: string;
  exam_id: string;
}